uploader/uploader/requirements.txt
coverage
requests==2.22.0
requests-file==1.4.3
requests-ftp==0.3.1
requests-mock==1.7.0
Pillow==6.2.1
VWS-Auth-Tools==2019.12.27.1
VWS-Python==2019.12.30.0
VWS-Python-Mock==2019.12.27.0